requirements.txt
streamlit==1.32.0
langchain==0.1.16
langchain-community==0.0.34
langchain-groq==0.1.3
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pypdf==4.2.0
python-dotenv==1.0.1
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2